Back to sun now, some new features just removed temporarily.
1. Back to sun now, some new features just removed temporarily.

+++ Deploy.sh
@@ -30,13 +30,6 @@ BLUE="\033[1;34m"
 OFF="\033[m"
 
 #
-# Define two status: 0 - Success, Turn on,
-#                    1 - Failure, Turn off
-#
-kBASHReturnSuccess=0
-kBASHReturnFailure=1
-
-#
 # Located repository.
 #
 REPO=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
@@ -74,8 +67,8 @@ t_clover_tools="/Volumes/EFI/EFI/CLOVER/tools"
 # Gvariables stands for getting datas from OS X.
 #
 gArgv=""
-gDebug=${kBASHReturnFailure}
-gProductVer=""
+gDebug=1
+gProductVersion=""
 target_website=""
 target_website_status=""
 RETURN_VAL=""
@@ -89,7 +82,7 @@ gVerticalRez=""
 gSystemRez=""
 gSystemHorizontalRez=""
 gSystemVerticalRez=""
-gPatchIOKit=${kBASHReturnSuccess}
+gPatchIOKit=0
 gClover_ig_platform_id=""
 target_ig_platform_id=""
 find_lid_byte_ENCODE=""
@@ -108,22 +101,10 @@ find_handoff_bytes=""
 replace_handoff_bytes=""
 find_handoff_bytes_ENCODE=""
 replace_handoff_bytes_ENCODE=""
-gTriggerLE=${kBASHReturnFailure}
-gProductVer="$(sw_vers -productVersion)"
-gOSVer=${gProductVer:0:5}
-gMINOR_VER=${gProductVer:3:2}
+gTriggerLE=1
+gMINOR_VER=$([[ "$(sw_vers -productVersion)" =~ [0-9]+\.([0-9]+) ]] && echo ${BASH_REMATCH[1]})
 gBak_Time=$(date +%Y-%m-%d-h%H_%M_%S)
 gBak_Dir="${REPO}/Backups/${gBak_Time}"
-gStop_Bak=${kBASHReturnFailure}
-gRecoveryHD=""
-gRecoveryHD_DMG="/Volumes/Recovery HD/com.apple.recovery.boot/BaseSystem.dmg"
-gTarget_rhd_Framework=""
-gTarget_Framework_Repo=""
-
-#
-# Set delimitation OS ver
-#
-let gDelimitation_OSVer=12
 
 #
 # Define target website
@@ -188,19 +169,10 @@ function _update()
 #--------------------------------------------------------------------------------
 #
 
-function _locate_rhd()
+function locate_esp()
 {
-    #
-    # Passing gRecoveryHD from ${targetEFI}
-    #
-    local gDisk_INF="$1"
-
-    #
-    # Example:
-    #
-    # disk0s3
-    # ^^^^^
-    diskutil list | grep -i "${gDisk_INF:0:5}" | grep "Recovery HD" |sed 's/.*MB   //'
+    diskutil info $1 | grep -i "Partition UUID" >${EFI_INFO}
+    targetUUID=$(grep -i "Disk / Partition UUID" ${EFI_INFO} | awk -F':' '{print $2}')
 }
 
 #
@@ -248,9 +220,9 @@ function _tidy_exec()
         #
         # Make the output clear.
         #
-        $1 >/tmp/report 2>&1 && RETURN_VAL=${kBASHReturnSuccess} || RETURN_VAL=${kBASHReturnFailure}
+        $1 >/tmp/report 2>&1 && RETURN_VAL=0 || RETURN_VAL=1
 
-        if [ "${RETURN_VAL}" == ${kBASHReturnSuccess} ];
+        if [ "${RETURN_VAL}" == 0 ];
           then
             _PRINT_MSG "OK: $2"
           else
@@ -350,16 +322,10 @@ function _getEDID()
         #
         # Get horizontal resolution. Arrays start from 0.
         #
-        # Examples:
-        #
-        # 00ffffffffffff004c2d240137314a4d0d1001036c221b782aaaa5a654549926145054bfef808180714f010101010101010101010101302a009851002a4030701300520e1100001e000000fd00384b1e510e000a202020202020000000fc0053796e634d61737465720a20
-        #                                                                                                                     ^
-        #                                                                                                                 ^^
-        #                                                                                                                           ^
-        #                                                                                                                       ^^
         gHorizontalRez_pr=${gEDID:116:1}
         gHorizontalRez_st=${gEDID:112:2}
         gHorizontalRez=$((0x$gHorizontalRez_pr$gHorizontalRez_st))
+
         #
         # Get vertical resolution. Actually, Vertical rez is no more needed in this scenario, but we just use this to make the
         # progress clear.
@@ -386,41 +352,18 @@ function _getEDID()
         #
         # Note: the argument of gPatchIOKit is set to 0 as default if the examination of resolution fail, this argument can ensure all models being powered up.
         #
-        gPatchIOKit=${kBASHReturnSuccess}
+        gPatchIOKit=0
       else
         #
         # No, patch IOKit is not required, we won't touch IOKit(for a more intergration/clean system since less is more).
         #
-        gPatchIOKit=${kBASHReturnFailure}
+        gPatchIOKit=1
     fi
 
     #
     # Passing gPatchIOKit to gPatchRecoveryHD.
     #
     gPatchRecoveryHD=${gPatchIOKit}
-}
-
-#
-#--------------------------------------------------------------------------------
-#
-
-function _unlock_pixel_clock()
-{
-    if [ $gMINOR_VER -ge $gDelimitation_OSVer ];
-      then
-        #
-        # 10.12+
-        #
-        gTarget_rhd_Framework="$gMountPoint/System/Library/Frameworks/CoreDisplay.framework/Versions/Current/CoreDisplay"
-      else
-        #
-        # 10.12-
-        #
-        gTarget_rhd_Framework="$gMountPoint/System/Library/Frameworks/IOKit.framework/Versions/Current/IOKit"
-    fi
-
-    sudo perl -i.bak -pe 's|\xB8\x01\x00\x00\x00\xF6\xC1\x01\x0F\x85|\x33\xC0\x90\x90\x90\x90\x90\x90\x90\xE9|sg' ${gTarget_rhd_Framework}
-    _tidy_exec "sudo codesign -f -s - ${gTarget_rhd_Framework}" "Patch and sign framework for Recovery HD"
 }
 
 #
@@ -478,20 +421,8 @@ function _check_and_fix_config()
             sed -ig "s/$gClover_ig_platform_id/$target_ig_platform_id/g" ${config_plist}
         fi
     fi
-    #
-    # Check SSDT-m-M3800.aml
-    #
-    local dCheck_SSDT="SSDT-m-M3800.aml"
-    local gSortedOrder=$(awk '/<key>SortedOrder<\/key>.*/,/<\/array>/' ${config_plist} | egrep -o '(<string>.*</string>)' | sed -e 's/<\/*string>//g')
-    local gSortedNumber=$(awk '/<key>SortedOrder<\/key>.*/,/<\/array>/' ${config_plist} | egrep -o '(<string>.*</string>)' | sed -e 's/<\/*string>//g' | wc -l)
-    if [[ $gSortedOrder != *"$dCheck_SSDT" ]];
-      then
-        #
-        # $dCheck_SSDT no found. Insert it.
-        #
-        /usr/libexec/plistbuddy -c "Add ':ACPI:SortedOrder:' string" ${config_plist}
-        /usr/libexec/plistbuddy -c "Set ':ACPI:SortedOrder:$gSortedNumber' $dCheck_SSDT" ${config_plist}
-    fi
+
+
 
     #
     # Repair the lid wake problem for 0x0a2e0008 by syscl/lighting/Yating Zhou.
@@ -726,14 +657,19 @@ function _find_acpi()
 
 function _update_clover()
 {
-    KEXT_DIR=/Volumes/EFI/EFI/CLOVER/kexts/${gOSVer}
+    #
+    # Gain OS generation.
+    #
+    gProductVersion="$(sw_vers -productVersion)"
+    OS_Version=$(echo ${gProductVersion:0:5})
+    KEXT_DIR=/Volumes/EFI/EFI/CLOVER/kexts/${OS_Version}
 
     #
     # Updating kexts. NOTE: This progress will remove any previous kexts.
     #
     _PRINT_MSG "--->: ${BLUE}Updating kexts...${OFF}"
     _tidy_exec "rm -rf ${KEXT_DIR}" "Remove pervious kexts in ${KEXT_DIR}"
-    _tidy_exec "cp -R ./CLOVER/kexts/${gOSVer} /Volumes/EFI/EFI/CLOVER/kexts/" "Update kexts from ./CLOVER/kexts/${gOSVer}"
+    _tidy_exec "cp -R ./CLOVER/kexts/${OS_Version} /Volumes/EFI/EFI/CLOVER/kexts/" "Update kexts from ./CLOVER/kexts/${OS_Version}"
     _tidy_exec "cp -R ./Kexts/*.kext ${KEXT_DIR}/" "Update kexts from ./Kexts"
 
     #
@@ -946,7 +882,15 @@ function _del()
 {
     local target_file=$1
 
-    _tidy_exec "sudo rm -R ${target_file}" "Remove ${target_file}"
+    if [ -d ${target_file} ];
+      then
+        _tidy_exec "sudo rm -R ${target_file}" "Remove ${target_file}"
+      else
+        if [ -f ${target_file} ];
+          then
+            _tidy_exec "sudo rm ${target_file}" "Remove ${target_file}"
+        fi
+    fi
 }
 
 #
@@ -998,66 +942,6 @@ function _fix_usb_ejected_improperly()
 #--------------------------------------------------------------------------------
 #
 
-function _printBackupLOG()
-{
-    #
-    # Examples:
-    #
-    # 2016-07-17-h01_43_14
-    # ^^^^ ^^ ^^
-    #             ^^ ^^ ^^
-    local gDAY="${gBak_Time:5:2}/${gBak_Time:8:2}/${gBak_Time:0:4}"
-    local gTIME="${gBak_Time:12:2}:${gBak_Time:15:2}:${gBak_Time:18:2}"
-    local gBackupLOG=$(echo "${gBak_Dir}/BackupLOG.txt")
-
-    #
-    # Print Header.
-    #
-    echo "  Backup Recovery HD(BaseSystem.dmg)"                                                   > "${gBackupLOG}"
-    echo ''                                                                                       >>"${gBackupLOG}"
-    echo "  DATE:                     $gDAY"                                                      >>"${gBackupLOG}"
-    echo "  TIME:                     $gTIME"                                                     >>"${gBackupLOG}"
-    local gRecentFileMD5=$(md5 -q "${gBak_BaseSystem}")
-    echo "  Origin Recovery HD MD5:   ${gRecentFileMD5}"                                          >>"${gBackupLOG}"
-    local gPatchedFileMD5=$(md5 -q "${gBaseSystem_PATCH}")
-    echo "  Patched Recovery HD MD5:  ${gPatchedFileMD5}"                                         >>"${gBackupLOG}"
-    echo ''                                                                                       >>"${gBackupLOG}"
-}
-
-#
-#--------------------------------------------------------------------------------
-#
-
-function _bakBaseSystem()
-{
-    gLastOpenedFileMD5=$(md5 -q "${gRecoveryHD_DMG}")
-
-    if [ -d "${REPO}/Backups" ];
-      then
-        if [[ `ls ${REPO}/Backups/*` == *'.txt'* ]];
-          then
-            gBakFileNames=($(ls ${REPO}/Backups/*/*.txt))
-        fi
-    fi
-
-    if [[ "${#gBakFileNames[@]}" -gt 0 ]];
-      then
-        gBakFileMD5=($(cat ${gBakFileNames[@]} | grep 'Patched Recovery HD MD5:' | sed -e 's/.*: //' -e 's/ //'))
-        for checksum in "${gBakFileMD5[@]}"
-        do
-          if [[ $checksum == $gLastOpenedFileMD5 && gStop_Bak != ${kBASHReturnSuccess} ]];
-            then
-              _PRINT_MSG "OK: Backup found. No more patch operations need"
-              gStop_Bak=${kBASHReturnSuccess}
-          fi
-        done
-    fi
-}
-
-#
-#--------------------------------------------------------------------------------
-#
-
 function _recoveryhd_fix()
 {
     #
@@ -1090,14 +974,14 @@ function _recoveryhd_fix()
     #
     # Mount Recovery HD.
     #
+    local gRecoveryHD=""
     local gMountPoint="/tmp/RecoveryHD"
     local gBaseSystem_RW="/tmp/BaseSystem_RW.dmg"
+    local gRecoveryHD_DMG="/Volumes/Recovery HD/com.apple.recovery.boot/BaseSystem.dmg"
     local gBaseSystem_PATCH="/tmp/BaseSystem_PATCHED.dmg"
-
-    #
-    # Locate Recovery HD
-    #
-    gRecoveryHD=$(_locate_rhd ${targetEFI})
+    diskutil list
+    printf "Enter ${RED}Recovery HD's ${OFF}IDENTIFIER, e.g. ${BOLD}disk0s3${OFF}"
+    read -p ": " gRecoveryHD
     _tidy_exec "diskutil mount ${gRecoveryHD}" "Mount ${gRecoveryHD}"
     _touch "${gMountPoint}"
 
@@ -1108,13 +992,12 @@ function _recoveryhd_fix()
     local gTarget_FS=$(echo 'UDRW')
 
     #
-    # Backup origin BaseSystem.dmg to ${REPO}/Backups
+    # Backup origin BaseSystem.dmg to ${REPO}/Backups/.
     #
-    _bakBaseSystem
     _touch "${gBak_Dir}"
-    cp "${gRecoveryHD_DMG}" "${gBak_Dir}"
-    gBak_BaseSystem="${gBak_Dir}/BaseSystem.dmg"
-    chflags nohidden "${gBak_BaseSystem}"
+    cp "${gRecoveryHD_DMG}" "${gBak_Dir}/"
+    local gBak_BaseSystem="${gBak_Dir}/BaseSystem.dmg"
+    _tidy_exec "chflags nohidden "${gBak_BaseSystem}"" "Show the dmg"
 
     #
     # Start to override.
@@ -1122,7 +1005,8 @@ function _recoveryhd_fix()
     _PRINT_MSG "--->: Convert ${gBaseSystem_FS}(r/o) to ${gTarget_FS}(r/w) ..."
     _tidy_exec "hdiutil convert "${gBak_BaseSystem}" -format ${gTarget_FS} -o ${gBaseSystem_RW} -quiet" "Convert ${gBaseSystem_FS}(r/o) to ${gTarget_FS}(r/w)"
     _tidy_exec "hdiutil attach "${gBaseSystem_RW}" -nobrowse -quiet -readwrite -noverify -mountpoint ${gMountPoint}" "Attach Recovery HD"
-    _unlock_pixel_clock
+    sudo perl -i.bak -pe 's|\xB8\x01\x00\x00\x00\xF6\xC1\x01\x0F\x85|\x33\xC0\x90\x90\x90\x90\x90\x90\x90\xE9|sg' $gMountPoint/System/Library/Frameworks/IOKit.framework/Versions/Current/IOKit
+    _tidy_exec "sudo codesign -f -s - $gMountPoint/System/Library/Frameworks/IOKit.framework/Versions/Current/IOKit" "Sign IOKit for Recovery HD"
     _tidy_exec "hdiutil detach $gMountPoint" "Detach mountpoint"
     #
     # Convert to origin format.
@@ -1185,6 +1069,7 @@ function main()
     diskutil list
     printf "Enter ${RED}EFI's${OFF} IDENTIFIER, e.g. ${BOLD}disk0s1${OFF}"
     read -p ": " targetEFI
+    locate_esp ${targetEFI}
     _tidy_exec "diskutil mount ${targetEFI}" "Mount ${targetEFI}"
 
     #
@@ -1316,11 +1201,6 @@ function main()
         _tidy_exec "cp "${prepare}"/CpuPm-4712HQ.aml "${compile}"/SSDT-pr.aml" "Generate C-States and P-State for Intel ${BLUE}i7-4712HQ${OFF}"
     fi
 
-    if [[ `sysctl machdep.cpu.brand_string` == *"i5-4200H"* ]]
-      then
-        _tidy_exec "cp "${prepare}"/CpuPm-4200H.aml "${compile}"/SSDT-pr.aml" "Generate C-States and P-State for Intel ${BLUE}i5-4200H${OFF}"
-    fi
-
     #
     # Install SSDT-m for ALS0.
     #
@@ -1365,20 +1245,8 @@ function main()
         # Patch IOKit.
         #
         _PRINT_MSG "--->: ${BLUE}Unlocking maximum pixel clock...${OFF}"
-        if [ $gMINOR_VER -ge $gDelimitation_OSVer ];
-          then
-            #
-            # 10.12+
-            #
-            gTarget_Framework_Repo="/System/Library/Frameworks/CoreDisplay.framework/Versions/Current/CoreDisplay"
-          else
-            #
-            # 10.12-
-            #
-            gTarget_Framework_Repo="/System/Library/Frameworks/IOKit.framework/Versions/Current/IOKit"
-        fi
-            sudo perl -i.bak -pe 's|\xB8\x01\x00\x00\x00\xF6\xC1\x01\x0F\x85|\x33\xC0\x90\x90\x90\x90\x90\x90\x90\xE9|sg' ${gTarget_Framework_Repo}
-            _tidy_exec "sudo codesign -f -s - ${gTarget_Framework_Repo}" "Patch and sign framework"
+        sudo perl -i.bak -pe 's|\xB8\x01\x00\x00\x00\xF6\xC1\x01\x0F\x85|\x33\xC0\x90\x90\x90\x90\x90\x90\x90\xE9|sg' /System/Library/Frameworks/IOKit.framework/Versions/Current/IOKit
+        _tidy_exec "sudo codesign -f -s - /System/Library/Frameworks/IOKit.framework/Versions/Current/IOKit" "Sign IOKit"
     fi
 
     #
